Add build to deploy.yml, Change from src/ to dist/ folders.
An earlier iteration of this site didn't require building as it was
based on static HTML so I removed the NodeJS installation and yarn
build steps and deployed only the src/ folder.

The current version of the site requires a build step to optimize
resources, and its output is the dist/ folder. This will ensure
that the appropriate files get pushed to AWS.