Remove unnecessary dependencies
The template I used as a base includes dozens of alternative pages
and layouts, including various JS libraries to show off functionality.
For the end site I've selected, almost all of this is no longer
useful, so we eliminate it to cut down on deployment time and to
save money on the storage bill.