Add Gilfoyle Lambda Slack Bot - Initial Commit
There is a burning need for a Slack Bot which responds with cutting
remarks from our favorite illegal alien. This commit establishes the
framework for providing this service by creating a python lambda
to handle the requests and an API Gateway to receive them.

Incidentally, I'm terrible at encapsulation, so we get a very minor
refactor of main.tf into site-* tf files to better separate concerns.
Apologies for the messy commit. It'll probably happen again.