Produce a functional Gilfoyle Slack Bot
This creates an API Gateway Endpoint -> Lambda integration using
Python 3.8 and boto3/botocore. I store messages and a pointer in
DynamoDB and retrieve those on invocation. This prevents me from
repeating a message too soon (the pool of viable insults is only
27 entries). On each request, we increment the pointer and fire a
Slack Webhook to the channel the user was in when invoking the
Slash Command.